requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0
openai>=1.35.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
jinja2>=3.1.2
